requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SQLAlchemy>=2.5
SQLAlchemy>=1.4
marshmallow>=3.0
pymssql>=2.2
python-dotenv>=0.21 
Flask-RESTX>=1.1.0 
flasgger
fastapi
apispec
apispec_webframeworks
flask-swagger-ui
dependency-injector==4.41.0
flask-socketio>=5.3.2
eventlet>=0.33.0
bcrypt
pyjwt
google-generativeai